more organizing

+++ src/actor_player.c
@@ -1,13 +1,12 @@
 #include "tools.h"
 #include "assets.h"
 #include "actor_factory.h"
-#include "actor_player.h"
 
 // private header
 void player_preload_assets(struct Actor* actor);
 void player_actor_update(struct Actor* actor);
-void player_actor_drawworld(struct Actor* actor, float delta_time);
-void player_actor_drawhud(struct Actor* actor, float delta_time);
+void player_actor_drawworld(struct Actor* actor, double delta_time);
+void player_actor_drawhud(struct Actor* actor, double delta_time);
 void player_actor_destroy(struct Actor* actor);
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -41,23 +40,23 @@ void player_actor_update(struct Actor* actor)
 	const ply_speed = 10;
 	if (actor->position.x > boundary)
 	{
-		actor->velocity.x = -ply_speed;
-		actor->velocity.y = RAND_RANGE(-ply_speed, ply_speed);
+		actor->velocity.x = (float)-ply_speed;
+		actor->velocity.y = (float)RAND_RANGE(-ply_speed, ply_speed);
 	}
 	if (actor->position.y > boundary)
 	{
-		actor->velocity.x = RAND_RANGE(-ply_speed, ply_speed);
-		actor->velocity.y = -ply_speed;
+		actor->velocity.x = (float)RAND_RANGE(-ply_speed, ply_speed);
+		actor->velocity.y = (float)-ply_speed;
 	}
 	if (actor->position.x < 0)
 	{
-		actor->velocity.x = +ply_speed;
-		actor->velocity.y = RAND_RANGE(-ply_speed, ply_speed);
+		actor->velocity.x = (float)+ply_speed;
+		actor->velocity.y = (float)RAND_RANGE(-ply_speed, ply_speed);
 	}
 	if (actor->position.y < 0)
 	{
-		actor->velocity.x = RAND_RANGE(-ply_speed, ply_speed);
-		actor->velocity.y = +ply_speed;
+		actor->velocity.x = (float)RAND_RANGE(-ply_speed, ply_speed);
+		actor->velocity.y = (float)+ply_speed;
 	}
 	
 	//actor->rotation = QuaternionMultiply(actor->rotation, QuaternionFromEuler(15 * DEG2RAD, 0, 0));
@@ -65,13 +64,13 @@ void player_actor_update(struct Actor* actor)
 
 void player_actor_drawworld(struct Actor* actor, double tick_percent)
 {
-	DrawCube(Vector3Add(Vector3Scale(VEC3FORWARD,5.0f), Vector3Scale(ACTOR_POS_DELTA(actor, tick_percent), 0.01f)), 0.2f, 0.2f, 0.2f, RED);
+	DrawCube(Vector3Add(Vector3Scale(VEC3FORWARD,5.0f), Vector3Scale(ACTOR_POS_DELTA(actor, (float)tick_percent), 0.01f)), 0.2f, 0.2f, 0.2f, RED);
 }
 
 void player_actor_drawhud(struct Actor* actor, double tick_percent)
 {
-	Vector3 delta_pos = ACTOR_POS_DELTA(actor, tick_percent);
-	DrawTextureEx(AssetGet_Texture(PLAYER_ASSET_TEXTURE), (Vector2){ delta_pos.x, delta_pos.y }, QuaternionToEuler(ACTOR_ROT_DELTA(actor, tick_percent)).x * RAD2DEG, 1, WHITE);
+	Vector3 delta_pos = ACTOR_POS_DELTA(actor, (float)tick_percent);
+	DrawTextureEx(AssetGet_Texture(PLAYER_ASSET_TEXTURE), (Vector2){ delta_pos.x, delta_pos.y }, QuaternionToEuler(ACTOR_ROT_DELTA(actor, (float)tick_percent)).x * RAD2DEG, 1, WHITE);
 }
 
 void player_actor_destroy(struct Actor* actor)
